Fix import cycle between problem.py and passes/definitions.py
Moved Format, ParseError, and DumpError from passes/definitions.py to
problem.py, since these are fundamental abstractions used by
ReductionProblem.view(). The definitions module now re-exports them
for backward compatibility.

This breaks the cycle:
- Before: definitions -> problem -> definitions
- After: definitions -> problem (no reverse dependency)
