+++ notes/import-graph.md
@@ -57,7 +57,6 @@ flowchart TD
     n13 --> n10
     n13 --> n14
     n14 --> n3
-    n14 --> n7
     n14 --> n24
     n16 --> n5
     n16 --> n6
