+++ src/shrinkray/passes/definitions.py
@@ -10,14 +10,32 @@ This module defines the core type aliases and abstractions for reduction:
 These abstractions enable format-agnostic reduction: the same pass
 (e.g., "delete duplicate elements") can work on bytes, lines, tokens,
 JSON arrays, or any other sequence-like type.
+
+Note: Format, ParseError, and DumpError are defined in shrinkray.problem
+and re-exported here for backward compatibility.
 """
 
-from abc import ABC, abstractmethod
 from collections.abc import Awaitable, Callable
 from functools import wraps
 from typing import TypeVar
 
-from shrinkray.problem import ReductionProblem
+from shrinkray.problem import (
+    DumpError,
+    Format,
+    ParseError,
+    ReductionProblem,
+)
+
+
+# Re-export for backward compatibility
+__all__ = [
+    "DumpError",
+    "Format",
+    "ParseError",
+    "ReductionPass",
+    "ReductionPump",
+    "compose",
+]
 
 
 S = TypeVar("S")
@@ -34,70 +52,6 @@ ReductionPass = Callable[[ReductionProblem[T]], Awaitable[None]]
 # The reducer runs passes on the pumped result using backtrack() to try to
 # reduce it below the original size.
 ReductionPump = Callable[[ReductionProblem[T]], Awaitable[T]]
-
-
-class ParseError(Exception):
-    """Raised when a Format cannot parse its input."""
-
-    pass
-
-
-class DumpError(Exception):
-    """Raised when a Format cannot serialize its output.
-
-    This occurs because not all internal representations map to valid
-    output in the target format. For example, a reduction might create
-    an invalid AST structure that cannot be converted back to source code.
-    """
-
-    pass
-
-
-class Format[S, T](ABC):
-    """A bidirectional transformation between two types.
-
-    Formats enable format-agnostic passes by abstracting the
-    parse/serialize cycle. For example:
-
-    - Split(b"\\n"): bytes <-> list[bytes] (lines)
-    - Tokenize(): bytes <-> list[bytes] (tokens)
-    - JSON: bytes <-> Any (Python objects)
-    - DimacsCNF: bytes <-> list[list[int]] (SAT clauses)
-
-    A Format must satisfy the round-trip property:
-        dumps(parse(x)) should be equivalent to x
-        (possibly with normalization)
-
-    Example usage:
-        # Delete duplicate lines
-        compose(Split(b"\\n"), delete_duplicates)
-
-        # Reduce integer literals in source code
-        compose(IntegerFormat(), reduce_integer)
-    """
-
-    @property
-    def name(self) -> str:
-        """Human-readable name for this format, used in pass names."""
-        return repr(self)
-
-    @abstractmethod
-    def parse(self, input: S) -> T:
-        """Parse input into the target type. Raises ParseError on failure."""
-        ...
-
-    def is_valid(self, input: S) -> bool:
-        """Check if input can be parsed by this format."""
-        try:
-            self.parse(input)
-            return True
-        except ParseError:
-            return False
-
-    @abstractmethod
-    def dumps(self, input: T) -> S:
-        """Serialize the target type back to the source type."""
-        ...
 
 
 def compose(format: Format[S, T], reduction_pass: ReductionPass[T]) -> ReductionPass[S]:
